refactor: extract memory cli module
